SE REALIZO UNA REFACTORIZACIÓN DE LOS PAQUETES PARA APLICAR EL MVC.
