+++ src/Imagen/lapiz.png
