+++ src/VISTA/frmProfesores.java
@@ -3,7 +3,7 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package proyectocs;
+package VISTA;
 
 import java.awt.Color;
 import javax.swing.JOptionPane;
